backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
websockets==13.1
python-dotenv==1.0.1
python-multipart==0.0.12
pytest==8.3.3
pytest-asyncio==0.24.0

# Real broker SDKs
kiteconnect>=5.0.1
fyers-apiv3>=3.0.0
SmartApi>=1.4.0
upstox-python-sdk>=2.9.0
dhanhq>=2.0.0
